Squish type ID and GC state boxed fields to i8
typeId was orignally i16 because it was intended to be unique for every
record type. gcState was i16 just to take advantage of the extra space
given by alignment.

The current plan is to have a separate recordClassId field that only
exists for record types. This allows us to squish these fields down to
i8. This leaves more room with stuffing small fields and flags at the
beginning of a boxed type without overflowing our 24 byte Cons.

+++ runtime/alloc/GarbageState.h
@@ -6,7 +6,7 @@
 namespace lliby
 {
 
-enum class GarbageState : std::uint16_t
+enum class GarbageState : std::uint8_t
 {
 	GlobalConstant = 0,
 	StackAllocation = 1,
